Add cedarldap directory and associated files
