+++ .envrc
@@ -60,7 +60,10 @@ PATH_add ./bin
 export CEDAR_ENV=dev
 export CEDAR_API_URL="webmethods-apigw.cedardev.cms.gov"
 export CEDAR_DIRECTORY=pkg/cedar
-export CEDAR_SWAGGER_FILE=$CEDAR_DIRECTORY/swagger-$CEDAR_ENV.json
+export CEDAR_EASI_DIRECTORY=$CEDAR_DIRECTORY/cedareasi
+export CEDAR_LDAP_DIRECTORY=$CEDAR_DIRECTORY/cedarldap
+export CEDAR_EASI_SWAGGER_FILE=$CEDAR_EASI_DIRECTORY/swagger-$CEDAR_ENV.json
+export CEDAR_LDAP_SWAGGER_FILE=$CEDAR_LDAP_DIRECTORY/swagger-$CEDAR_ENV.json
 # Load a local overrides file. Any changes you want to make for your local
 # environment should live in that file.
