+++ pkg/cedar/cedareasi/cedar_test.go
@@ -1,4 +1,4 @@
-package cedar
+package cedareasi
 
 import (
 	"testing"
