+++ pkg/cedar/cedareasi/gen/client/operations/intakegovernance_p_o_s_t4_responses.go
@@ -13,7 +13,7 @@ import (
 
 	strfmt "github.com/go-openapi/strfmt"
 
-	"github.com/cmsgov/easi-app/pkg/cedar/gen/models"
+	models "github.com/cmsgov/easi-app/pkg/cedar/cedareasi/gen/models"
 )
 
 // IntakegovernancePOST4Reader is a Reader for the IntakegovernancePOST4 structure.
