+++ pkg/cedar/cedareasi/gen/client/operations/systemid_g_e_t3_responses.go
@@ -13,7 +13,7 @@ import (
 
 	strfmt "github.com/go-openapi/strfmt"
 
-	"github.com/cmsgov/easi-app/pkg/cedar/gen/models"
+	models "github.com/cmsgov/easi-app/pkg/cedar/cedareasi/gen/models"
 )
 
 // SystemidGET3Reader is a Reader for the SystemidGET3 structure.
