+++ pkg/cedar/cedareasi/gen/client/operations/systems_g_e_t1_responses.go
@@ -13,7 +13,7 @@ import (
 
 	strfmt "github.com/go-openapi/strfmt"
 
-	"github.com/cmsgov/easi-app/pkg/cedar/gen/models"
+	models "github.com/cmsgov/easi-app/pkg/cedar/cedareasi/gen/models"
 )
 
 // SystemsGET1Reader is a Reader for the SystemsGET1 structure.
