+++ pkg/cedar/cedareasi/gen/client/operations/systemsid_g_e_t2_responses.go
@@ -13,7 +13,7 @@ import (
 
 	strfmt "github.com/go-openapi/strfmt"
 
-	"github.com/cmsgov/easi-app/pkg/cedar/gen/models"
+	models "github.com/cmsgov/easi-app/pkg/cedar/cedareasi/gen/models"
 )
 
 // SystemsidGET2Reader is a Reader for the SystemsidGET2 structure.
