+++ pkg/cedar/cedareasi/gen/models/software_product.go
@@ -7,6 +7,7 @@ package models
 
 import (
 	strfmt "github.com/go-openapi/strfmt"
+
 	"github.com/go-openapi/swag"
 )
