+++ pkg/cedar/cedareasi/translated_client_test.go
@@ -1,4 +1,4 @@
-package cedar
+package cedareasi
 
 import (
 	"fmt"
